Created the logic to display available profiles to invite.
Created the ProfileManager class and added a function called
profiles_available_to_invite() which will fetch and return all the
profiles which are neither in the friends list nor do they have any
pending connection/invitation requests with the current logged in user.

Modified views.py and created a template to display the list which is
fetched in the above logic.